Add Promotion Container component
