+++ nekoyume/Assets/AddressableAssets/UI/Module/Lobby/RaidBtn.prefab
@@ -64869,6 +64869,7 @@ GameObject:
   - component: {fileID: 134801035865149620}
   - component: {fileID: 1081515367264110028}
   - component: {fileID: 7100399749330656870}
+  - component: {fileID: 6609180674501105850}
   m_Layer: 5
   m_Name: Promotion
   m_TagString: Untagged
@@ -64973,7 +64974,20 @@ MonoBehaviour:
     sprite: {fileID: 21300000, guid: 9f36daf9a31f3cc41994d461a36468db, type: 3}
   - languageType: 3
     sprite: {fileID: 21300000, guid: 622b4c4427d62c84cbe1f5c233a4d7ac, type: 3}
+--- !u!114 &6609180674501105850
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6736566957761785786}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 606a65fd6eca47988dc65a11d28abc30, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   season: 2
+  times: []
 --- !u!1 &6878589271596072809
 GameObject:
   m_ObjectHideFlags: 0
