+++ nekoyume/Assets/_Scripts/UI/Module/PromotionContainer.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 606a65fd6eca47988dc65a11d28abc30
+timeCreated: 1702457234
